+++ jkube-kit/config/service/src/test/java/org/eclipse/jkube/kit/config/service/ApplyServiceTest.java
@@ -27,8 +27,10 @@ import io.fabric8.kubernetes.api.model.ReplicationController;
 import io.fabric8.kubernetes.api.model.ReplicationControllerBuilder;
 import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.ServiceBuilder;
+import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.CustomResourceDefinitionListBuilder;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.apps.DeploymentBuilder;
+import io.fabric8.kubernetes.client.utils.Serialization;
 import io.fabric8.openshift.api.model.DeploymentConfigBuilder;
 import io.fabric8.openshift.api.model.Project;
 import io.fabric8.openshift.api.model.ProjectBuilder;
@@ -36,6 +38,7 @@ import io.fabric8.openshift.api.model.Route;
 import io.fabric8.openshift.api.model.RouteBuilder;
 import io.fabric8.openshift.client.server.mock.OpenShiftServer;
 import mockit.Mocked;
+import org.eclipse.jkube.kit.common.GenericCustomResource;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.config.service.openshift.WebServerEventCollector;
 import org.junit.Before;
@@ -44,7 +47,6 @@ import org.junit.Test;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 import static java.net.HttpURLConnection.HTTP_CREATED;
@@ -67,6 +69,31 @@ public class ApplyServiceTest {
     public final OpenShiftServer mockServer = new OpenShiftServer(false);
 
     private ApplyService applyService;
+
+    private final CustomResourceDefinition virtualServiceCRD = new CustomResourceDefinitionBuilder()
+            .withNewMetadata().withName("networking.istio.io.virtualservices").endMetadata()
+            .withNewSpec()
+            .withGroup("networking.istio.io")
+            .withScope("Namespaced")
+            .withVersion("v1alpha3")
+            .withNewNames()
+            .withKind("VirtualService")
+            .withPlural("virtualservices")
+            .endNames()
+            .endSpec()
+            .build();
+    private final CustomResourceDefinition gatewayCRD = new CustomResourceDefinitionBuilder()
+            .withNewMetadata().withName("networking.istio.io.gateway").endMetadata()
+            .withNewSpec()
+            .withGroup("networking.istio.io")
+            .withScope("Namespaced")
+            .withVersion("v1alpha3")
+            .withNewNames()
+            .withKind("Gateway")
+            .withPlural("gateways")
+            .endNames()
+            .endSpec()
+            .build();
 
     @Before
     public void setUp() {
@@ -112,7 +139,7 @@ public class ApplyServiceTest {
                 .once();
 
         // When
-        applyService.applyEntities(fileName, entities, log, 5, null, null, null);
+        applyService.applyEntities(fileName, entities, log, 5);
 
         // Then
         collector.assertEventsRecordedInOrder("new-rc", "new-configmap", "new-service", "new-deploy", "new-pod");
@@ -201,21 +228,17 @@ public class ApplyServiceTest {
     }
 
     @Test
-    public void testProcessCustomEntities() throws Exception {
+    public void testApplyGenericCustomResource() throws Exception {
         // Given
-        List<String> crdNames = new ArrayList<>();
-        crdNames.add("virtualservices.networking.istio.io");
-        crdNames.add("gateways.networking.istio.io");
-        File crFragmentDir = new File(getClass().getResource("/gateway-cr.yml").getFile()).getParentFile();
+        File gatewayFragment = new File(getClass().getResource("/gateway-cr.yml").getFile());
+        File virtualServiceFragment = new File(getClass().getResource("/virtualservice-cr.yml").getFile());
+        GenericCustomResource gateway = Serialization.yamlMapper().readValue(gatewayFragment, GenericCustomResource.class);
+        GenericCustomResource virtualService = Serialization.yamlMapper().readValue(virtualServiceFragment, GenericCustomResource.class);
         WebServerEventCollector collector = new WebServerEventCollector();
         mockServer.expect().get()
-                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/virtualservices.networking.istio.io")
-                .andReply(collector.record("get-crd-virtualservice").andReturn(HTTP_OK, getVirtualServiceIstioCRD()))
-                .always();
-        mockServer.expect().get()
-                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/gateways.networking.istio.io")
-                .andReply(collector.record("get-crd-gateway").andReturn(HTTP_OK, getGatewayIstioCRD()))
-                .always();
+                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions")
+                .andReply(collector.record("get-crds").andReturn(HTTP_OK, new CustomResourceDefinitionListBuilder().withItems(virtualServiceCRD, gatewayCRD).build()))
+                .times(2);
         mockServer.expect().post()
                 .withPath("/apis/networking.istio.io/v1alpha3/namespaces/default/virtualservices")
                 .andReply(collector.record("post-cr-virtualservice").andReturn(HTTP_OK, "{}"))
@@ -226,70 +249,64 @@ public class ApplyServiceTest {
                 .once();
 
         // When
-        applyService.applyCustomEntities(crdNames, crFragmentDir, null, Collections.emptyList());
+        applyService.applyGenericCustomResource(gateway, gatewayFragment.getName());
+        applyService.applyGenericCustomResource(virtualService, virtualServiceFragment.getName());
 
         // Then
-        collector.assertEventsRecordedInOrder("get-crd-virtualservice", "get-crd-gateway", "post-cr-virtualservice", "post-cr-gateway");
+        collector.assertEventsRecordedInOrder("get-crds", "post-cr-gateway", "get-crds", "post-cr-virtualservice");
         assertEquals(7, mockServer.getMockServer().getRequestCount());
     }
 
     @Test
     public void testProcessCustomEntitiesReplaceCustomResources() throws Exception {
         // Given
-        List<String> crdNames = new ArrayList<>();
-        crdNames.add("virtualservices.networking.istio.io");
-        crdNames.add("gateways.networking.istio.io");
-        File crFragmentDir = new File(getClass().getResource("/gateway-cr.yml").getFile()).getParentFile();
+        File gatewayFragment = new File(getClass().getResource("/gateway-cr.yml").getFile());
+        File virtualServiceFragment = new File(getClass().getResource("/virtualservice-cr.yml").getFile());
+        GenericCustomResource gateway = Serialization.yamlMapper().readValue(gatewayFragment, GenericCustomResource.class);
+        GenericCustomResource virtualService = Serialization.yamlMapper().readValue(virtualServiceFragment, GenericCustomResource.class);
         WebServerEventCollector collector = new WebServerEventCollector();
         mockServer.expect().get()
-                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/virtualservices.networking.istio.io")
-                .andReply(collector.record("get-crd-virtualservice").andReturn(HTTP_OK, getVirtualServiceIstioCRD()))
-                .always();
-        mockServer.expect().get()
-                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/gateways.networking.istio.io")
-                .andReply(collector.record("get-crd-gateway").andReturn(HTTP_OK, getGatewayIstioCRD()))
-                .always();
+                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions")
+                .andReply(collector.record("get-crds").andReturn(HTTP_OK, new CustomResourceDefinitionListBuilder().withItems(virtualServiceCRD, gatewayCRD).build()))
+                .times(2);
         mockServer.expect().get()
                 .withPath("/apis/networking.istio.io/v1alpha3/namespaces/default/virtualservices/reviews-route")
-                .andReply(collector.record("get-cr-virtualservice").andReturn(HTTP_OK, "{}"))
-                .once();
+                .andReply(collector.record("get-cr-virtualservice").andReturn(HTTP_OK, "{\"metadata\":{\"resourceVersion\":\"1001\"}}"))
+                .times(2);
         mockServer.expect().put()
                 .withPath("/apis/networking.istio.io/v1alpha3/namespaces/default/virtualservices/reviews-route")
                 .andReply(collector.record("put-cr-virtualservice").andReturn(HTTP_OK, "{}"))
                 .once();
         mockServer.expect().get()
                 .withPath("/apis/networking.istio.io/v1alpha3/namespaces/default/gateways/mygateway-https")
-                .andReply(collector.record("get-cr-gateway").andReturn(HTTP_OK, "{}"))
-                .once();
+                .andReply(collector.record("get-cr-gateway").andReturn(HTTP_OK, "{\"metadata\":{\"resourceVersion\":\"1002\"}}"))
+                .times(2);
         mockServer.expect().put()
                 .withPath("/apis/networking.istio.io/v1alpha3/namespaces/default/gateways/mygateway-https")
                 .andReply(collector.record("put-cr-gateway").andReturn(HTTP_OK, "{}"))
                 .once();
 
         // When
-        applyService.applyCustomEntities(crdNames, crFragmentDir, null, Collections.emptyList());
+        applyService.applyGenericCustomResource(gateway, gatewayFragment.getName());
+        applyService.applyGenericCustomResource(virtualService, virtualServiceFragment.getName());
 
         // Then
-        collector.assertEventsRecordedInOrder("get-crd-virtualservice", "get-crd-gateway", "get-cr-virtualservice", "put-cr-virtualservice", "get-cr-gateway", "put-cr-gateway");
-        assertEquals(7, mockServer.getMockServer().getRequestCount());
+        collector.assertEventsRecordedInOrder("get-crds", "get-cr-gateway", "get-cr-gateway", "put-cr-gateway", "get-crds", "get-cr-virtualservice", "get-cr-virtualservice", "put-cr-virtualservice");
+        assertEquals(9, mockServer.getMockServer().getRequestCount());
     }
 
     @Test
     public void testProcessCustomEntitiesRecreateModeTrue() throws Exception {
         // Given
-        List<String> crdNames = new ArrayList<>();
-        crdNames.add("virtualservices.networking.istio.io");
-        crdNames.add("gateways.networking.istio.io");
-        File crFragmentDir = new File(getClass().getResource("/gateway-cr.yml").getFile()).getParentFile();
+        File gatewayFragment = new File(getClass().getResource("/gateway-cr.yml").getFile());
+        File virtualServiceFragment = new File(getClass().getResource("/virtualservice-cr.yml").getFile());
+        GenericCustomResource gateway = Serialization.yamlMapper().readValue(gatewayFragment, GenericCustomResource.class);
+        GenericCustomResource virtualService = Serialization.yamlMapper().readValue(virtualServiceFragment, GenericCustomResource.class);
         WebServerEventCollector collector = new WebServerEventCollector();
         mockServer.expect().get()
-                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/virtualservices.networking.istio.io")
-                .andReply(collector.record("get-crd-virtualservice").andReturn(HTTP_OK, getVirtualServiceIstioCRD()))
-                .always();
-        mockServer.expect().get()
-                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/gateways.networking.istio.io")
-                .andReply(collector.record("get-crd-gateway").andReturn(HTTP_OK, getGatewayIstioCRD()))
-                .always();
+                .withPath("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions")
+                .andReply(collector.record("get-crds").andReturn(HTTP_OK, new CustomResourceDefinitionListBuilder().withItems(virtualServiceCRD, gatewayCRD).build()))
+                .times(2);
         mockServer.expect().delete()
                 .withPath("/apis/networking.istio.io/v1alpha3/namespaces/default/virtualservices/reviews-route")
                 .andReply(collector.record("delete-cr-virtualservice").andReturn(HTTP_OK, "{}"))
@@ -309,25 +326,13 @@ public class ApplyServiceTest {
         applyService.setRecreateMode(true);
 
         // When
-        applyService.applyCustomEntities(crdNames, crFragmentDir, null, Collections.emptyList());
-        applyService.setRecreateMode(false);
+        applyService.applyGenericCustomResource(gateway, gatewayFragment.getName());
+        applyService.applyGenericCustomResource(virtualService, virtualServiceFragment.getName());
 
         // Then
-        collector.assertEventsRecordedInOrder("get-crd-virtualservice", "get-crd-gateway",
-                "delete-cr-virtualservice", "post-cr-virtualservice", "delete-cr-gateway", "post-cr-gateway");
+        collector.assertEventsRecordedInOrder("get-crds", "delete-cr-gateway", "post-cr-gateway", "get-crds", "delete-cr-virtualservice", "post-cr-virtualservice");
         assertEquals(7, mockServer.getMockServer().getRequestCount());
-    }
-
-    @Test
-    public void testProcessCustomEntitiesWithNullCustomResourceList() throws Exception {
-        // Given
-        File crFragmentDir = new File(getClass().getResource("/gateway-cr.yml").getFile()).getParentFile();
-
-        // When
-        applyService.applyCustomEntities(null, crFragmentDir, null, Collections.emptyList());
-
-        // Then
-        assertEquals(1, mockServer.getMockServer().getRequestCount()); // This one is just /apis check
+        applyService.setRecreateMode(false);
     }
 
     @Test