+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Usage:
 * Fix #509: Port of ServiceDiscoveryEnricher from FMP
 * Fix #510: Update Compatibility matrix for OpenShift 4.5 and 4.6
 * Fix #511: Namespace as resource fragment results in NullPointerException
+* Fix #515: Properties now get resolved in CustomResource fragments
 * Fix #521: NPE on Buildconfig#getContextDir if `<dockerFile>` references a file with no directory
 * Fix #513: openshift-maven-plugin: service.yml fragment with ports creates service with unnamed port mapping
 * Fix #231: IngressEnricher ignores IngressRules defined in XML config